+++ pyflwdir/__init__.py
@@ -2,7 +2,7 @@ import pkg_resources
 
 # submodules
 from pyflwdir.flwdir import *
-from pyflwdir.flwdir_raster import *
+from pyflwdir.pyflwdir import *
 from pyflwdir.dem import *
 
 from pyflwdir import gis_utils